cleanup usage of TeachingPopoverCarouselContextValues
