+++ packages/react-components/react-teaching-popover-preview/src/components/TeachingPopoverCarousel/index.ts
@@ -3,4 +3,4 @@ export * from './TeachingPopoverCarousel.types';
 export * from './renderTeachingPopoverCarousel';
 export * from './useTeachingPopoverCarousel';
 export * from './useTeachingPopoverCarouselStyles.styles';
-export * from './TeachingPopoverCarouselContext';
+export * from './useTeachingPopoverCarouselContextValues';
