+++ packages/react-components/react-teaching-popover-preview/src/components/TeachingPopoverCarousel/useTeachingPopoverCarouselContextValues.ts
@@ -1,0 +1,19 @@
+import type {
+  TeachingPopoverCarouselContextValues,
+  TeachingPopoverCarouselState,
+} from './TeachingPopoverCarousel.types';
+
+export function useTeachingPopoverCarouselContextValues_unstable(
+  state: TeachingPopoverCarouselState,
+): TeachingPopoverCarouselContextValues {
+  const { store, value, setValue, onPageChange, carouselWalker } = state;
+
+  const carousel = {
+    store,
+    value,
+    setValue,
+    onPageChange,
+  };
+
+  return { carousel, carouselWalker };
+}
